LayoutTests/http/tests/websocket/tests/hybi/nocache-expected.txt

Test whether WebSocket request contains no-cache headers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
PASS message is "PASS"
onclose() was called.
PASS closeEvent.wasClean is true
PASS successfullyParsed is true

TEST COMPLETE
